[refactor] 삭제 시 비밀번호 입력
원하는 id의 행을 삭제할떄 비밀번호도 입력하도록함. 만약 비번이 맞으면 정상삭제되고, 틀리면 에러 반환

+++ src/main/java/com/example/todoapp/repository/JdbcTemplateTodoRepository.java
@@ -7,7 +7,6 @@ import org.springframework.jdbc.core.JdbcTemplate;
 
 import org.springframework.stereotype.Repository;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -29,9 +28,10 @@ public class JdbcTemplateTodoRepository implements TodoRepository{
 
     @Override
     public List<ResponseDto> findAllTodo() {
-        String sql = "SELECT userName, todo, doDate, createDate, updateDate FROM todo";
+        String sql = "SELECT id, userName, todo, doDate, createDate, updateDate FROM todo ORDER BY updateDate Desc ";
         return jdbcTemplate.query(sql, (rs, rowNum) ->
                 new ResponseDto(
+                        rs.getLong("id"),
                         rs.getString("userName"),
                         rs.getString("todo"),
                         rs.getDate("doDate"),
@@ -46,6 +46,7 @@ public class JdbcTemplateTodoRepository implements TodoRepository{
         String sql = "SELECT * FROM todo WHERE id = ?";
         return jdbcTemplate.queryForObject(sql, (rs, rowNum) ->
                 new ResponseDto(
+                        rs.getLong("id"),
                         rs.getString("userName"),
                         rs.getString("todo"),
                         rs.getDate("doDate"),
@@ -72,12 +73,18 @@ public class JdbcTemplateTodoRepository implements TodoRepository{
 
     @Override
     public boolean deleteTodoById(RequestDto dto) {
-        String sql = "DELETE FROM todo WHERE id = ?";
-        if (dto.getId() != null) {
-            jdbcTemplate.update(sql, dto.getId());
-            return true;
-        } else {
+        String findPw = "SELECT password FROM todo WHERE id =?";
+        String Pw = jdbcTemplate.queryForObject(findPw, new Object[]{dto.getId()}, String.class);
+        if (!dto.getPassword().equals(Pw)) {
             return false;
+        } else {
+            if (dto.getId() != null) {
+                String sql = "DELETE FROM todo WHERE id = ?";
+                jdbcTemplate.update(sql, dto.getId());
+                return true;
+            } else {
+                return false;
+            }
         }
     }
